backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
openai>=0.28.1
python-dotenv==1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
sqlalchemy>=1.4.0
pandas>=1.5.0
openpyxl>=3.0.0

